Move classes player/genres in file player.py, move display_logo
